refactor(decompose): split webapp/electron.vite.config.ts into folder
